Added AJAX chat example
Added simple AJAX chat example to /src/php/ajaxDemo.
You have to enter the contact phonenumber in index.php, and your
credentials in socket.php.
Start the demo by going to index.php.
This will start a socket listener process in the background. This
process will pull and push messages to and from the $_SESSION variable.
The queues for inbound and outbound messages are stored in the $_SESSION
global variable.
At this time only sending and receiving text messages to and from a
single contact is supported.
I plan to add contact picture, composing status and message delivery
status to this example. If there is enough interest I might also create
a new repository focused on this :)
Now excuse me while I'll try to get some sleep, been up all night, it's
6:21 AM :")