Added SendSetRecoveryToken, (read description!)
SendSetRecoveryToken is kind of a big deal here. It allows you to set
the recovery token for your account.
This means that you can use that token in checkCredentials($token) to
request a new password at any time!
This also means that you can now use that cursed Login() method to log
into your account using the recovery token.
I also had to update checkCredentials to work properly.

My next change will be to replace the IMEI with the recovery token in
the constructor so that you no longer need to provide a custom token
when using Login() (currently woring as Login($token))