NO CODE CHANGES. Formatting only.
This commit begins the attempt to give all the code in this project some consistent formatting. We are *attempting* to try and bring the code into alignment with PSR-2 ( http://www.php-fig.org/psr/2/ ) but there are still many inconsistencies.

There is no wish to start any "standards" war here. If there is agreement we can code to any other formatting/coding standard in the future.

Comments welcome.